+++ Gems/Atom/Tools/ShaderManagementConsole/Code/Source/Platform/Linux/ShaderManagementConsole_Traits_Linux.h
@@ -7,5 +7,3 @@
  */
 #pragma once
 
-#define AZ_TRAIT_SHADER_MANAGEMENT_CONSOLE_EXT ""
-
