+++ Gems/Atom/Tools/ShaderManagementConsole/Code/Source/Platform/Mac/ShaderManagementConsole_Traits_Mac.h
@@ -7,5 +7,3 @@
  */
 #pragma once
 
-#define AZ_TRAIT_SHADER_MANAGEMENT_CONSOLE_EXT ".app"
-
