+++ Gems/Atom/Tools/ShaderManagementConsole/Code/Source/Platform/Windows/ShaderManagementConsole_Traits_Windows.h
@@ -7,5 +7,3 @@
  */
 #pragma once
 
-#define AZ_TRAIT_SHADER_MANAGEMENT_CONSOLE_EXT ".exe"
-
