+++ Gems/Atom/Tools/ShaderManagementConsole/Code/Source/ShaderManagementConsoleApplication.h
@@ -8,12 +8,21 @@
 
 #pragma once
 
+#include <Atom/RPI.Reflect/Material/MaterialAsset.h>
 #include <AtomToolsFramework/Document/AtomToolsDocumentApplication.h>
+#include <AtomToolsFramework/Window/AtomToolsMainWindowFactoryRequestBus.h>
+#include <AzToolsFramework/API/EditorWindowRequestBus.h>
+#include <ShaderManagementConsoleRequestBus.h>
+#include <Window/ShaderManagementConsoleBrowserInteractions.h>
+#include <Window/ShaderManagementConsoleWindow.h>
 
 namespace ShaderManagementConsole
 {
     class ShaderManagementConsoleApplication
         : public AtomToolsFramework::AtomToolsDocumentApplication
+        , private ShaderManagementConsoleRequestBus::Handler
+        , private AtomToolsFramework::AtomToolsMainWindowFactoryRequestBus::Handler
+        , private AzToolsFramework::EditorWindowRequestBus::Handler
     {
     public:
         AZ_TYPE_INFO(ShaderManagementConsole::ShaderManagementConsoleApplication, "{A31B1AEB-4DA3-49CD-884A-CC998FF7546F}");
@@ -21,13 +30,31 @@ namespace ShaderManagementConsole
         using Base = AtomToolsFramework::AtomToolsDocumentApplication;
 
         ShaderManagementConsoleApplication(int* argc, char*** argv);
+        ~ShaderManagementConsoleApplication();
 
         // AzFramework::Application overrides...
-        void CreateStaticModules(AZStd::vector<AZ::Module*>& outModules) override;
+        void Reflect(AZ::ReflectContext* context) override;
         const char* GetCurrentConfigurationName() const override;
+        void StartCommon(AZ::Entity* systemEntity) override;
 
         // AtomToolsFramework::AtomToolsApplication overrides...
         AZStd::string GetBuildTargetName() const override;
         AZStd::vector<AZStd::string> GetCriticalAssetFilters() const override;
+
+        // AzToolsFramework::EditorWindowRequests::Bus::Handler
+        QWidget* GetAppMainWindow() override;
+
+        // AtomToolsMainWindowFactoryRequestBus::Handler overrides...
+        void CreateMainWindow() override;
+        void DestroyMainWindow() override;
+
+        // ShaderManagementConsoleRequestBus::Handler overrides...
+        AZ::Data::AssetInfo GetSourceAssetInfo(const AZStd::string& sourceAssetFileName) override;
+        AZStd::vector<AZ::Data::AssetId> FindMaterialAssetsUsingShader(const AZStd::string& shaderFilePath) override;
+        AZStd::vector<AZ::RPI::ShaderCollection::Item> GetMaterialInstanceShaderItems(const AZ::Data::AssetId& assetId) override;
+
+    private:
+        AZStd::unique_ptr<ShaderManagementConsoleWindow> m_window;
+        AZStd::unique_ptr<ShaderManagementConsoleBrowserInteractions> m_assetBrowserInteractions;
     };
 } // namespace ShaderManagementConsole
