+++ Gems/Atom/Tools/ShaderManagementConsole/Code/Source/Window/ShaderManagementConsoleWindow.h
@@ -14,7 +14,6 @@
 #include <AtomToolsFramework/Document/AtomToolsDocumentMainWindow.h>
 
 AZ_PUSH_DISABLE_WARNING(4251 4800, "-Wunknown-warning-option") // disable warnings spawned by QT
-#include <Window/ToolBar/ShaderManagementConsoleToolBar.h>
 #include <QStandardItemModel>
 AZ_POP_DISABLE_WARNING
 #endif
@@ -36,8 +35,7 @@ namespace ShaderManagementConsole
         ~ShaderManagementConsoleWindow() = default;
 
     protected:
+        bool GetOpenDocumentParams(AZStd::string& openPath) override;
         QWidget* CreateDocumentTabView(const AZ::Uuid& documentId) override;
-
-        ShaderManagementConsoleToolBar* m_toolBar = {};
     };
 } // namespace ShaderManagementConsole
