--- Gems/Atom/Tools/ShaderManagementConsole/Code/shadermanagementconsoledocument_files.cmake
@@ -1,17 +1,0 @@
-#
-# Copyright (c) Contributors to the Open 3D Engine Project.
-# For complete copyright and license terms please see the LICENSE at the root of this distribution.
-#
-# SPDX-License-Identifier: Apache-2.0 OR MIT
-#
-#
-
-set(FILES
-    Include/Atom/Document/ShaderManagementConsoleDocumentModule.h
-    Include/Atom/Document/ShaderManagementConsoleDocumentRequestBus.h
-    Source/Document/ShaderManagementConsoleDocument.cpp
-    Source/Document/ShaderManagementConsoleDocument.h
-    Source/Document/ShaderManagementConsoleDocumentModule.cpp
-    Source/Document/ShaderManagementConsoleDocumentSystemComponent.cpp
-    Source/Document/ShaderManagementConsoleDocumentSystemComponent.h
-)
