--- Gems/Atom/Tools/ShaderManagementConsole/Code/shadermanagementconsolewindow_files.cmake
@@ -1,22 +1,0 @@
-#
-# Copyright (c) Contributors to the Open 3D Engine Project.
-# For complete copyright and license terms please see the LICENSE at the root of this distribution.
-#
-# SPDX-License-Identifier: Apache-2.0 OR MIT
-#
-#
-
-set(FILES
-    Include/Atom/Window/ShaderManagementConsoleWindowModule.h
-    Include/Atom/Core/ShaderManagementConsoleRequestBus.h
-    Source/Window/ShaderManagementConsoleBrowserInteractions.h
-    Source/Window/ShaderManagementConsoleBrowserInteractions.cpp
-    Source/Window/ShaderManagementConsoleWindow.h
-    Source/Window/ShaderManagementConsoleWindow.cpp
-    Source/Window/ShaderManagementConsoleWindowModule.cpp
-    Source/Window/ShaderManagementConsole.qrc
-    Source/Window/ShaderManagementConsoleWindowComponent.h
-    Source/Window/ShaderManagementConsoleWindowComponent.cpp
-    Source/Window/ToolBar/ShaderManagementConsoleToolBar.h
-    Source/Window/ToolBar/ShaderManagementConsoleToolBar.cpp
-)
